refactor: move index::util to a separate file
